+++ src/JsonDecoder/ChunkDecoder.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace JsonMachine\JsonDecoder;
+
+
+interface ChunkDecoder
+{
+    public function decodeInternalKey($jsonScalarKey): ChunkDecodingStringResult;
+
+    public function decodeKey($jsonScalarKey): ChunkDecodingResult;
+
+    public function decodeValue($jsonValue): ChunkDecodingResult;
+}